feat: add `Copy` bound to `LogItem` feat: read file with bincode
opt: appender

fix: ub and style
fix: writer UB

Because implementation relies on `LogItem` being bare data that can be read bytes by bytes, this bound force the invariant so that no extra data can be held (i.e. ptrs)
chore

style: reformat and fix clippy issue

doc: add architecture diagram

+++ src/checker/log/index_checker.rs
@@ -1,8 +1,8 @@
 use std::mem::size_of;
 
-use crate::formats::log::IndexFileHeader;
-use crate::util::BlockIODevice;
 use anyhow::{anyhow, ensure, Result};
+
+use crate::{formats::log::IndexFileHeader, util::BlockIODevice};
 
 pub(crate) struct IndexChecker<'a, F> {
     file: &'a mut F,
@@ -34,7 +34,7 @@ impl<'a, F: BlockIODevice> IndexChecker<'a, F> {
             "Invalid log file: broken header"
         );
         ensure!(
-            (*self.file_size - size_of::<IndexFileHeader>() as u64) % self.index_size as u64 == 0,
+            (*self.file_size - size_of::<IndexFileHeader>() as u64) % self.index_size == 0,
             "Broken index file"
         );
